changing appearance of checkboxes of contracts list

+++ src/main/java/com/endava/service_system/controller/rest/InvoicesRestController.java
@@ -212,7 +212,7 @@ public class InvoicesRestController {
 
     @GetMapping("/company/invoices/allIds")
     public ResponseEntity getAllCompanyInvoicesId(){
-        int[] invoicesIds = invoiceService.getAllInvoicesIdsbyCompanyUsername(authUtils.getAuthenticatedUsername());
+        Long[] invoicesIds = invoiceService.getAllInvoicesIdsbyCompanyUsername(authUtils.getAuthenticatedUsername());
         LOGGER.log(Level.DEBUG, invoicesIds);
         return new ResponseEntity(invoicesIds, HttpStatus.OK);
     }
